make property construction/assignment more clear
It is currently not easy to understand which constructors and
assignment operators for properties are actually used for certain
arguments.

This can be made more clear by using individual overloads for the
different uses cases, including construction and assignment from other
properties.

Note that this comes at the cost of minor code redundancy.