remove harmful property conversion operators
Properties can currently implicitly be converted to a reference to
their value.

While this is convenient in a variety of use cases, it also comes with
a number of pitfalls that can result in unexpected behavior.

Most notably, expressions might result in the unintended usage of
certain operators, because the property can be converted to a
corresponding argument. This is particularly common with integer
properties, which might be converted in a variety of unexpected ways.

Since these conversions depend on the value type, it is also often not
obvious why certain expressions work and others don't, which can result
in confusion and inconsistent code.

Additionally, the conversions add further complexity to all operations
involving both properties and raw values (e.g. for serialization),
where extra care is required to prevent unintentional usage of the
wrong overload.

For these reasons, the potential hazards outweigh the benefits and the
operators should rather be removed entirely.

Note that this is a breaking change, which can be accommodated by
explicitly acquiring the property value (e.g. via the value operator).