simplify property interface
Properties currently explicitly distinguish between construction and
assignment of values and require the property to be in a corresponding
state before the operation.

This was originally introduced to provide an additional layer of
safety, by forcing users to specify the expected state.

However, in practice this distinction is usually not relevant and comes
at the cost of a convoluted and overengineered interface consisting of
way more functions than necessary.

Through conceding this distinction, the interface can be simplified
substantially by merging the construction and assignment functions.

Note that this is a breaking change that can be accommodated for by
replacing the removed functions with 'emplace'.