improve debugging of static properties
The storage of static properties is currently implemented by using an
std::aligned_storage.

While this is a simple solution for raw aligned memory, it does not
provide any information about the stored value, which severely impedes
debugging and analysis.

Instead, the storage can be made value-based, by using an anonymous
union and a few helper types.

Note that the implementation is loosely based on what is often used for
std::optional and similar to the solution that had to be removed in an
earlier commit (see [1]).

References:

 - [1] commit 91ce6af8750263a1c16c4676ff7553de07801393