fix list constructor of boolean vector properties
Properties of boolean vectors can currently no longer be constructed
from initializer lists.

This is regression that was introduced when the initializer constructor
was converted into a non-template function (see [1]).

The issue occurs, because boolean vectors are implemented with a value
type of uint8_t instead of bool to avoid the std::vector<bool>
specialization, which is not compatible with the DOTS vector class.

References:

 - [1] commit 31294e4edd8bbe3b95b13a923955371b01d12d56