make property interface 'optional-like'
Since properties are conceptually close to std::optional, it would be
beneficial to provide a similar interface, with which many users might
already be familiar with.

Doing so requires the introduction of an 'invalid' type and a
corresponding constant (akin to std::nullopt), which can be used for
assignment and comparison.

Note that this also renames the 'destroy' function to 'reset' for
consistency.