improve error output for property value arguments
Value arguments for properties (e.g. for construction or assignment)
are currently always forwarded to the corresponding function of the
descriptor.

While this results in a static assertion error when the value type of
the property cannot be constructed from the arguments, it still usually
comes with a slew of error messages, which might obfuscate the actual
cause of the problem.

Instead, the arguments can be safeguarded via SFINAE at the topmost
level, which prevents the functions from being selected entirely.

Note that SFINAE is preferable, because it usually provides immediate
feedback in many IDEs, while a static assertion is only triggered at
build time.