resolve 'maybe uninitialized' warning with GCC 1x
Building the project with various GCC versions has shown that the
recently suppressed warning for uninitialized values with GCC 9 (see
[1]) can also occur in later GCC versions (albeit sometimes in
different manifestations).

Since there are currently no indications that the memory can actually
be uninitialized, the warnings should generally be suppressed when
building the project with GCC.

Note that the suppression is deliberately restricted to the property
types.

References:

 - [1] commit 14e5201825aa11f8cbbb93b9d382d267a3050d65