safeguard required C++ features
Since compilers more and more move in the direction of supporting many,
but not all features of a standard (particularly since C++20, see [1]),
it becomes challenging to ensure that a certain set of compilers is
supported.

Furthermore, users might get confusing build errors when using a
compiler that supports the version configured via CMake (e.g. C++20),
but is missing a required feature.

For this reason, the library should use features selectively and have a
mechanism that conveys clear errors to the user when the compiler is
incompatible. Unfortunately this cannot be done via CMake, which
discontinued support for handling particular low level C++ features
since C++17 (see [2]).

Another way to achieve this is to use feature testing macros, which
were introduced in C++20 (see [3]).

References:

 - [1] https://en.cppreference.com/w/cpp/compiler_support

 - [2] https://cmake.org/cmake/help/v3.20/prop_gbl/CMAKE_CXX_KNOWN_FEATURES.html#low-level-individual-compile-features

 - [3] https://en.cppreference.com/w/cpp/feature_test