update RapidJSON to latest master revision
The referenced version of RapidJSON is four years old and is starting
to show issues when being used with more recent C++ versions and
compiler combinations.

Since the library stopped releasing (tagged) versions a long time ago,
despite still being updated, the only option is to choose an arbitrary
revision, which might as well be the latest master version.