resolve Clang issue due to unsupported warning
Building the project with Clang currently results in warnings, due to
an unsupported flag.

This is a regression that was introduced while fixing an issue with
various versions of the GCC compiler (see [1]).

Since Clang supports most features and options of GCC, it also defines
the '__GNUG__' macro in addition to '__clang__'.

While this works in most cases, the '-Wmaybe-uninitialized' flag does
not seem to be supported by Clang (at least in version 10).

References:

 - [1] commit d56469d695714a191f1d3044dbcc3ca910896a0a