use designated struct initialization as default
The experimental designated initialization of DOTS structs has been
used extensively in various projects and has proven to be both viable
and stable.

Since the project is now based on C++20 anyway, there is no reason to
not make designated initialization the default.

Note that because it is not clear whether or not mixing initialization
approaches is safe, the initialization based on the PropertyInitializer
class is removed entirely instead of being deprecated.

Also note that the property types require a minor adjustment to allow
construction and assignment from arithmetic literals without producing
conversion warnings.