simplify suppression of initializer warnings
When using designated initializers, both GCC and Clang issue a warning
when initializers are missing for one or more members.

Since this is a common use case for DOTS structs, the warning is
currently suppressed globally.

While it does not seem possible to avoid the warning for DOTS structs
only, the suppression flag can be moved into a pragma in the generated
header files, which at least simplifies suppression of the warning for
both the library itself, as well as user applications.