use generated types for property tests
Tests for properties and proxy properties are currently based on a
custom property area and corresponding property types. This was
historically done to decouple the tests from the implementation of the
static struct types.

However, similarly to the static struct types itself (see [1]), this
approach is harmful, because the properties are not the static ones
that are actually used in real life applications.

Additionally, the current approach introduces additional complexity to
the tests, making them harder to understand an maintain.

References:

 - [1] commit 96b36cb9ef1d4ecc9342b297c92d13e3b631ef01