overhaul tests for property comparison operators
The property tests for both equality and ordering comparisons are
currently hard to understand and maintain, because they are scattered
in way too many test cases with confusing titles.

They are also incomplete and do not cover several use cases, such as
comparison with left hand values or 'dots::invalid'.

To rectify this, the tests can be overhauled completely for readability
and completeness.

Note that this refactoring reveals several bugs regarding the property
comparisons, that will be fixed in later commits for clarity.

Also note that the corresponding proxy property tests are redundant and
will be removed entirely to simplify maintenance.