fix ordering of invalid properties
When the properties were overhauled some time ago, the less operation
was defined in such a way, that invalid properties would always compare
as less than the smallest valid value.

While this was explicitly stated in the commit message (see [1]), it
was for some reason implemented in the exact opposite way, which is
also still reflected in the tests.

To make matters worse, comparisons with 'dots::invalid' were
implemented with the original assumption in mind (see [2]), making the
operator not only wrong but also inconsistent.

This commit implements the behavior consistently that was originally
intended.

References:

 - [1] commit ad6d6c6b799be946eea5ce2dd311de56b11f2d47

 - [2] commit 4ed94fe9f22d54dd7002404cdfa5251934793013