introduce version specific default TCP ports
Historically, only '11234' has been used as a default port for TCP
channels when none was specified.

This did not change when the message format was overhauled (see [1]),
resulting in potential issues, when mixing applications using different
versions (particularly current and legacy ones).

To be backwards compatible, the library should rather use default ports
that are specific to a particular message format version.

Note that this is a breaking change, as tcp-v1 (aka the "legacy
format") is now again mapped to '11234', while tcp-v2 channels (or tcp
respectively) use '11235'.

References:

 - [1] commit eef67e5ee8802148094154de749c6a7cc39235e5