move default TCP ports to channels and listeners
The default TCP ports are currently handled in the Application class,
which means that they are not applied when using transceivers directly
(e.g. in advanced utility applications).

Instead, the default TCP port handling should rather take place in the
TCP channels and listeners respectively, so that they are always used
when no port is specified via an endpoint.

Note that because listeners do not yet provide their endpoint via their
base class like channels do, this currently requires a workaround in
the host transceiver to include the default port in the log output when
listeners are added.