Channel: export types when sending trough a DescriptorData object
A out-of-order descriptor can be received when a program A sends a
DescriptorRequest to dotsd while simultaneously another program B is in
the process of sending its descriptors but program A missed dependent types.
Program A would try to register the new type and fail, because it could
not resolve the dependent types.
This change ensures, that even for this types the server sends the
dependent types into the stream to program A.