use header-only version of libfmt in unit tests
The unit tests are currently linking statically against libfmt, while
the library itself is using the header-only version.

Even though this might work with some compilers, it can result in
issues with others (e.g. MSVC), because symbols are potentially present
multiple times in several translation units.

Note that users of the library still have the option to use a static
version if their build setup supports it or if they are building DOTS
as a shared library.