address build issues with MSVC 19.40 and later
The project does not build with MSVC versions 19.40 and latest
(currently 14.42.34433), because certain syntactic constructs used in
the 'TestProperty' unit tests result in an 'internal compiler error'.

Further investigation showed that this seems to be related to a
compiler bug with current MSVC versions (see comment in unit test for
more detail).

Note that while the tests themselves can be guarded against newer MSVC
versions to make the project buildable again, we currently do not want
to officially declare the affected versions as supported, because it is
not clear what exactly the implications of the problem are.