fix invalid reads on serializer exceptions
The recently introduced serializer exceptions are currently reading
past the end of the input buffer when creating the hex dump, which
results in incorrect dumps at best and segfaults at worst.

This is caused by incorrectly constructing the span over the input
buffer, which requires the last element and not the element after the
last one like an end iterator (see [1]).

References:

 - [1] https://en.cppreference.com/w/cpp/container/span/span