increase coverage of unit test for CBOR skipping
While there already is a unit test for skipping values with a CBOR
reader, it does currently only cover a few select data types.

In particular, it does not cover the behavior for long byte arrays
(e.g. text strings), which was the reason the test was added in the
first place.